requirements.txt
Django==1.7.1
Pillow==2.6.1
djangorestframework==3.0.0
wsgiref==0.1.2
